Add optional SNI support via TxSNI library from pypi
Three configuration options are added to caldavd.plist to make this work:
* SNIMode (BOOL): Whether to use SNI. If not, falls back to old SSL code.
* SSLDir (STRING): Directory where all the key/cert/CAcert bundle .pem
  files live, indexed by hostname
* DefaultSSLBundle (STRING): What file to load as certificate when no
  hostname has been indicated by the client when connecting or no cert
  exists for the given host.

Certificate structure will be as such:
|_DirectoryNameIConfiguredInCaldavd.plist
  |_my.supercool.host.pem
  |_some.other.host.pem
  |_...

Symlinks also work if you have multiple domains covered on a single
cert.

CAVEATS: This mode of operation assumes that no certificates were
generated with keys that are protected by passphrases. The old
code which exists to account for this won't help you in SNI mode, as we
no longer go down those code paths (TxSNI doesn't appear to support
that, though with SNI it does sound somewhat hairy to have to deal with
for potentially a *lot* of certs).

No updates to the testsuite have been made for this change, but
considering that I didn't break the old pathways for SSL/TLS, it at
least doesn't break the current testsuite.